Split tsconfig.json into two files
It's due to VSCode's behavior. VSCode only scans configuration file with name tsconfig.json, not tsconfig.*.json, and this behavior is not configurable in IDE settings. Thus, to get full support from IDE regardless of whether we're editing the file that is included in tsconfig.json or not, we'll introduce tsconfig.prod.json that only describes which files are going to be included for final build.